requirements.txt
fastapi
uvicorn
python-dotenv
pyodbc
fastmcp
pydantic
